bot/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\notmm\blackjack\bot\bot.py: 14,15
APScheduler == 3.7.0

# C:\Users\notmm\blackjack\bot\bot.py: 10
# C:\Users\notmm\blackjack\bot\create_user.py: 2
# C:\Users\notmm\blackjack\bot\game\blackjack.py: 7,12
# C:\Users\notmm\blackjack\bot\misc.py: 7,14
# C:\Users\notmm\blackjack\bot\play.py: 3,4
# C:\Users\notmm\blackjack\bot\profile.py: 5
discord_py_slash_command == 2.4.1

# C:\Users\notmm\blackjack\bot\bot.py: 1,8,9
# C:\Users\notmm\blackjack\bot\create_user.py: 1
# C:\Users\notmm\blackjack\bot\misc.py: 3
# C:\Users\notmm\blackjack\bot\play.py: 1
# C:\Users\notmm\blackjack\bot\profile.py: 1,2
discord_timers == 0.1.0

# C:\Users\notmm\blackjack\bot\create_user.py: 4
# C:\Users\notmm\blackjack\bot\db.py: 1,2
pymongo[srv] == 3.12.0

# C:\Users\notmm\blackjack\bot\bot.py: 12
# C:\Users\notmm\blackjack\bot\db.py: 5
python_dotenv == 0.19.0

# C:\Users\notmm\blackjack\bot\bot.py: 6
rich == 10.6.0

dblpy==0.4.0